Gentleman/Gentleman/requirements.txt
asgiref==3.3.1
chardet==3.0.4
Django==3.1.3
django-newsletter==0.9.1
ldif3==3.1.1
Pillow==8.0.1
psycopg2==2.8.6
python-card-me==0.9.3
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.4.1
surlex==0.2.0
unicodecsv==0.14.1
